chore: remove circular dependency, rename intense to default in heading theme
