+++ packages/typography/src/components/text.tsx
@@ -1,11 +1,6 @@
 import styled, {css} from 'styled-components'
-import {getTextColor, getFontFamily} from '../utils'
 import {getTextStyle} from '../styles/text'
-
-export type TextVariant = 300 | 400 | 500 | 600
-export type TextFontFamily = 'ui' | 'display' | 'mono'
-export type TextColor = 'muted' | 'default' | 'intense'
-export type TextIntent = 'none' | 'success' | 'info' | 'danger' | 'warning'
+import {TextVariant, TextColor, TextIntent, TextFontFamily} from '../types'
 
 export interface TextProps {
   variant?: TextVariant
