+++ packages/typography/src/index.tsx
@@ -6,3 +6,4 @@ export {Label, LabelProps} from './components/label'
 export {getHeadingStyle} from './styles/headings'
 export {getParagraphStyle} from './styles/paragraph'
 export {getTextStyle} from './styles/text'
+export * from './types'
