+++ packages/typography/src/types.ts
@@ -1,0 +1,19 @@
+export type HeadingVariant = 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900
+export type HeadingFontFamily = 'ui' | 'display' | 'mono'
+
+export type ParagraphVariant = 300 | 400 | 500
+export type ParagraphColor = 'muted' | 'default' | 'intense'
+export type ParagraphFontFamily = 'ui' | 'display' | 'mono'
+
+export type TextVariant = 300 | 400 | 500 | 600
+export type TextFontFamily = 'ui' | 'display' | 'mono'
+export type TextColor = 'muted' | 'default' | 'intense'
+export type TextIntent = 'none' | 'success' | 'info' | 'danger' | 'warning'
+
+export type LabelVariant = 300 | 400 | 500
+export type LabelFontFamily = 'ui' | 'display' | 'mono'
+
+export type StrongVariant = 300 | 400 | 500 | 600
+export type StrongFontFamily = 'ui' | 'display' | 'mono'
+export type StrongColor = 'muted' | 'default' | 'intense'
+export type StrongIntent = 'none' | 'success' | 'info' | 'danger' | 'warning'
